refactor: Avoid Redis use in decrypt func
Now, the only responsability of Crypto is encrypting and decrypting. It
allows the library to be more scalable and don't have too many
responsability, such as, checking if the user has permissions to decrypt
the file or uploading the files to the DB. All of those responsabilities
must be implemented in other modules.