feature: Integrate tests with pre-commit and GitHub actions (#43)
* add: Add pre-commit config file

* add: Add github actions

* refactor: Include .github and .pre-commit-config

* add: Add pyproject and requirements

* refactor: Always run tests before commit

* fix: Fix pre-commit dependency

* refactor: Remove unused pytest.ini

* refactor: Fix redis service (#36)

- Pass file uuid to file fragment object.
- Capitalize first letter in doc string.

* refactor: Update api to use redis service (#45)

- Fix typo error
- Add some necessary blank spaces
- Delete unncessary comments

* feature: Add isort hook to pre-commit

* chore: Use poetry for dependency management

* fix: Order imports and remove trailing whitespaces

The imports have been ordered using isort, all the trailing spaces have
been removed and now all the files end with a new line.

* chore: Add ruff config to pyproject.toml

Ruff is now used as a linter for the python files.

* fix: Disable package mode

* fix: Install dependencies using --no-root flag

Make poetry to not try to install the project as a PyPI package, since
we only use poetry for dependencies management.

* chore: Mark FastAPI functions as immutable

Ruff was marking that calling fastapi methods such as fastapi.Depends or
fastapi.File in function definitions as an error. Since this calls
are intented and follow the practices of FastAPI, are now exceptioned.

* fix: Use exception chaining with `from`

Use exception chaining with `raise ... from` to preserve original
traceback.

* feature: Add ruff to pre-commit hooks

* feature: Add ruff to github actions ci

* chore: Run check-yaml hook in yml files only

* chore: Remove unnedeed blank line on module desc

---------

Co-authored-by: jorgerodriguez-palisis <jorge.rodriguez@palisis.com>