feature: Create login authentication (#49)
* chore: Add necessary dependencies to package.json

* chore: Modify pnpm lock

* chore: Add const and types file

- Add a file with some constants used in various file along the project.
- Type page, site and links.

* feat: Create toggle to handle theme functionality

Create theme toggle to change between dark and light theme, and manage
the icons related to each theme.

* feat: Create particles component

Component to add particles to the webpage background.

* feat: Create grid component

Create a component to build a grid that will be used on the webpage
background.

* feat: Create button component

Create component to use as Github link button.

* style: Add some icons for style

* feat: Add utils

Add utility related to tailwind merge and clsx inputs style.

* feat: Add a component for head page

* feat: Create header component

Create a header with logo, nav links, and theme toggle.

* feat: Create component for navigation bar

Add logic for render the navigation bar for different pages, except
login.

* feat: create login card component

Create a card with a form for manage the user credentials input.

* feat: Create login page

* feat: add head layout

Add head layout with common metadata, favicon, common links...

* feat: Create base layout

Create a layout that will work as base to all the pages. This layout
includes the head, grid and starry, header and navigation.

* style: Add global styles

Add few styles in order to apply them in all the pages.

* chore: Add end of line blank space

* chore: Add end of line on theme toggle

* fix: Delete unnecessary navbar tag on base layout

* style: Add new global styles

- Define few fonts.
- Define dark and light theme colors.

* chore: Establish dark theme by default

Add html tag to establish always the dark theme as theme by default.

* refactor: Normalize crypto library (#42)

* refactor: Normalize Crypto methods

Crypto object will now expose three methods: encrypt()
and decrypt(). This last one, now accepts fragmented and
non-fragmented data as a parameter.
Also, unit tests has been refactored to use the renamed Crypto's methods.

* fix: Fragment size corrected

The size of the fragment was using ** instead of *.

* refactor: Fix redis service (#36)

- Pass file uuid to file fragment object.
- Capitalize first letter in doc string.

* refactor: Update api to use redis service (#45)

- Fix typo error
- Add some necessary blank spaces
- Delete unncessary comments

* feature: Integrate tests with pre-commit and GitHub actions (#43)

* add: Add pre-commit config file

* add: Add github actions

* refactor: Include .github and .pre-commit-config

* add: Add pyproject and requirements

* refactor: Always run tests before commit

* fix: Fix pre-commit dependency

* refactor: Remove unused pytest.ini

* refactor: Fix redis service (#36)

- Pass file uuid to file fragment object.
- Capitalize first letter in doc string.

* refactor: Update api to use redis service (#45)

- Fix typo error
- Add some necessary blank spaces
- Delete unncessary comments

* feature: Add isort hook to pre-commit

* chore: Use poetry for dependency management

* fix: Order imports and remove trailing whitespaces

The imports have been ordered using isort, all the trailing spaces have
been removed and now all the files end with a new line.

* chore: Add ruff config to pyproject.toml

Ruff is now used as a linter for the python files.

* fix: Disable package mode

* fix: Install dependencies using --no-root flag

Make poetry to not try to install the project as a PyPI package, since
we only use poetry for dependencies management.

* chore: Mark FastAPI functions as immutable

Ruff was marking that calling fastapi methods such as fastapi.Depends or
fastapi.File in function definitions as an error. Since this calls
are intented and follow the practices of FastAPI, are now exceptioned.

* fix: Use exception chaining with `from`

Use exception chaining with `raise ... from` to preserve original
traceback.

* feature: Add ruff to pre-commit hooks

* feature: Add ruff to github actions ci

* chore: Run check-yaml hook in yml files only

* chore: Remove unnedeed blank line on module desc

---------

Co-authored-by: jorgerodriguez-palisis <jorge.rodriguez@palisis.com>

* chore: Add blank lines in try-except block

* refactor: Move types file into type directory

* fix: Fix types route

* feat: Create authentication context

Create method to authenticate user via JWT and handle the logout.

* feat: Add footer component

* feat: Add expand button component

* feat: Create protected route file

Create file to define what the protected routes are. In our case all the
routes are protected except login, so we define that here.

* feat: Add section component

* feat: Add dashboard page

Create main page of the app.

* refactor: Modify App.tsx

Modify app component to define protected routes and import different
pages.

* refactor: modify main entrypoint

Add authentication provider and browser router to app in entrypoint.

---------

Co-authored-by: Javi Correa <jcorreag@pm.me>