feature: Connect api with frontend (#75)
* chore: Install axios in frontend

* chore: Change config to work in local.

* feat: Implement get file method

Implement a method to get all the files that the user has stored in the
S3 bucket based on his user id.

* feat: Create list files endpoint

Create a endpoint to list the files that the user has stored.

* feat: Modify frontend to integrate API

Create several function to handle logic for upload, download and list
files in the frontend, this logic will be separate from the page in the
future.

* fix: Correct linter errors